Updating to used suggested Activity
